+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_08_18_000004) do
+ActiveRecord::Schema[8.1].define(version: 2026_08_19_000001) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
 
@@ -215,6 +215,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_08_18_000004) do
     t.string "est_label"
     t.text "feedback"
     t.datetime "finished_at"
+    t.string "pr_url"
     t.string "repo"
     t.boolean "risky", default: false, null: false
     t.datetime "started_at"